+++ app/src/main/res/layout/activity_exercise_detail.xml
@@ -171,7 +171,7 @@
     </TableLayout>
     </FrameLayout>
 
-    <com.example.dolbomi.LakuePagingButton
+    <com.example.dolbomi.paging.LakuePagingButton
         android:id="@+id/lpb_buttonlist"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
